Remove player entity
