--- carnage/players/vocations/knight.py
@@ -1,35 +1,0 @@
-# MIT License
-#
-# Copyright (c) 2022, Rodolfo Olivieri
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-"""Module that represents a Knight vocation."""
-from carnage.database.models.vocation.vocation import VocationModel
-from carnage.players.vocations.base import BaseVocation
-
-
-class Knight(BaseVocation):
-    """Class that overrides methods and properties of vocation."""
-
-    def __init__(self, vocation: VocationModel) -> None:
-        """Class that interprets an specific player.
-
-        :param vocation: The model that represents a vocation.
-        """
-        super().__init__(vocation)
