+++ carnage/vocations/base.py
