--- tests/unit_tests/players/player_test.py
@@ -1,6 +1,0 @@
-from carnage.players.player import Player
-
-
-def test_player_can_initialize():
-    player = Player()
-    assert isinstance(player, Player)
